Core Functionality done
The player can move around and jump,
The player can shoot, using object pooling system.
The enemies are spawning after a set time.
The enemies  are following the player and they shoot at the player if they are close enough.
If all the enemies are dead the spawn timer will reset.